backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["nwilson314 <nsw.wilson@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["standard"], version = "^0.115.8"}
sqlmodel = "^0.0.22"
alembic = "^1.14.1"
black = "^25.1.0"
loguru = "^0.7.3"
pydantic-settings = "^2.7.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
